refactor: update imports to use new structured model.cwm.resource.relational package
