+++ core/src/main/java/org/eclipse/daanse/rolap/element/RolapCatalog.java
@@ -112,8 +112,8 @@ import org.eclipse.daanse.rolap.common.member.SqlMemberSource;
 import org.eclipse.daanse.rolap.common.nativize.RolapNativeRegistry;
 import org.eclipse.daanse.rolap.common.star.RolapStar;
 import org.eclipse.daanse.rolap.common.star.RolapStarRegistry;
-import org.eclipse.daanse.cwm.util.objectmodel.core.Namespaces;
-import org.eclipse.daanse.cwm.util.resource.relational.ColumnSets;
+import org.eclipse.daanse.cwm.model.cwm.objectmodel.core.util.Namespaces;
+import org.eclipse.daanse.cwm.model.cwm.resource.relational.util.ColumnSets;
 import org.eclipse.daanse.rolap.mapping.model.database.relational.ColumnInternalDataType;
 import org.eclipse.daanse.rolap.util.ClassResolver;
 import org.slf4j.Logger;
