+++ core/src/main/java/org/eclipse/daanse/rolap/element/RolapLevel.java
@@ -38,7 +38,7 @@ import java.util.stream.Stream;
 import java.sql.Types;
 
 import org.eclipse.daanse.cwm.model.cwm.resource.relational.SQLSimpleType;
-import org.eclipse.daanse.cwm.util.resource.relational.SqlSimpleTypes;
+import org.eclipse.daanse.cwm.model.cwm.resource.relational.util.SqlSimpleTypes;
 import org.eclipse.daanse.sql.model.type.BestFitColumnType;
 import org.eclipse.daanse.sql.model.type.Datatype;
 import org.eclipse.daanse.olap.api.agg.Segment;
